CMSC-1283 - Newsletter: cast fix.
git-svn-id: https://scm.mmbase.org/mmbase/trunk@33377 91430c5f-d668-0410-9e7d-dd663366dcfd

+++ contributions/CMSContainer_Portlets/portlets-newsletter/src/java/com/finalist/newsletter/services/impl/NewsletterSubscriptionHibernateService.java
@@ -1,5 +1,6 @@
 package com.finalist.newsletter.services.impl;
 
+import java.math.BigInteger;
 import java.util.List;
 import java.util.Set;
 
@@ -83,7 +84,7 @@ public class NewsletterSubscriptionHibernateService extends HibernateService imp
    @Transactional
    public int getSubscribersRelatedInfoCount(Set<Long> authenticationIds, String fullName, String userName, String email, boolean paging) {
       Query query = executeSubscribersSearch(authenticationIds, fullName, userName, email, paging, true);
-      return (Integer)(query.uniqueResult());
+      return ((BigInteger)(query.uniqueResult())).intValue();
    }
    
 }
